feat: add admin endpoints for retrieving user session durations and get average session duration
