+++ src/route/admin.route.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import {adminLogin, forgotPassword, verifyOtp, resendOTP, resetPassword, updateAdminProfile, changePasswordAdmin} from "../controller/admin.controller.js"
+import {adminLogin, forgotPassword, verifyOtp, resendOTP, resetPassword, updateAdminProfile, changePasswordAdmin, getAverageSessionDuration, getUserSessionDurations} from "../controller/admin.controller.js"
 import {isAdmin} from "../middleware/role.middleware.js"
 import {verifyJWT} from "../middleware/auth.middleware.js";
 import upload from "../middleware/multer.middleware.js"
@@ -17,6 +17,11 @@ router.patch("/profile", verifyJWT, isAdmin, upload.single("avatar"), updateAdmi
 
 // Change Password Route
 router.patch("/change-password", verifyJWT, isAdmin, changePasswordAdmin);
+
+// Dashboard
+// Get session duration
+router.get("/session-durations/:userId",verifyJWT, isAdmin, getUserSessionDurations);
+router.get("/average-session-duration",verifyJWT, isAdmin, getAverageSessionDuration);
